pyproject.toml
[project]
name = "service-auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = []

[dependency-groups]
api = [
    "aiobotocore>=2.19.0",
    "fastapi>=0.115.8",
    "fastapi-libkit==0.1.3",
    "fastapi-sqlalchemy-toolkit>=0.7.11",
    "pydantic-settings>=2.7.1",
    "sqlmodel>=0.0.22",
    "ydb-async>=0.0.2",
]
